clarify example code comments
The comments in the example code made it sound like the world regions renderer was predefined _in_ the esri-leaflet-renderers.js file instead of at the service end point, which I could see confusing some people.

I've updated the comments so that they make more sense to me. Take it or leave it.